Bots/Obsidion/pyproject.toml
[tool.poetry]
authors = ["Darkflame72 <leon@bowie-co.nz>"]
description = "A Minecraft discord bot"
license = "AGPL-3.0-or-later"
name = "obsidion"
version = "2.0.0.dev1"

[tool.poetry.dependencies]
Babel = "^2.9.0"
aiodns = "^3.0.0"
aiohttp = "^3.6.1"
aioredis = "^2.0.0"
alembic = "^1.6.5"
asyncpg = "^0.24.0"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
lxml = "^4.5.2"
mcsrvstats = "^0.1.1"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
python = "^3.9"
uvloop = "^0.16.0"
SQLAlchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
beautifulsoup4 = "^4.9.3"
asyncpixel = "^1.1.1"
discord-py-slash-command = "^3.0.1"
dblpy = "^0.4.0"
sentry-sdk = "^1.3.1"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
Pygments = "^2.10.0"
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"
nox = "^2021.6.12"
nox-poetry = "^0.8.6"
pre-commit = "^2.13.0"
pre-commit-hooks = "^4.0.1"
reorder-python-imports = "^2.6.0"
safety = "^1.9.0"
flake8-bugbear = "^21.4.3"
pep8-naming = "^0.12.0"
flake8-builtins = "^1.5.3"
flake8-bandit = "^2.1.2"
flake8-eradicate = "^1.1.0"
flake8-comprehensions = "^3.6.1"
flake8-import-order = "^0.18.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.1"]
